refactor: Extract BaseLayerSettings
