+++ app/configurator/components/chart-options-selector.tsx
@@ -4,7 +4,7 @@ import { groups } from "d3-array";
 import get from "lodash/get";
 import groupBy from "lodash/groupBy";
 import dynamic from "next/dynamic";
-import { ReactNode, useCallback, useEffect, useMemo, useState } from "react";
+import { ReactNode, useCallback, useMemo, useState } from "react";
 
 import { DEFAULT_SORTING, getFieldComponentId } from "@/charts";
 import {
@@ -20,7 +20,6 @@ import {
   DEFAULT_FIXED_COLOR_FIELD_OPACITY,
   DEFAULT_OTHER_COLOR_FIELD_OPACITY,
 } from "@/charts/map/constants";
-import { getMap } from "@/charts/map/ref";
 import { useQueryFilters } from "@/charts/shared/chart-helpers";
 import { LegendItem, LegendSymbol } from "@/charts/shared/legend-color";
 import Flex from "@/components/flex";
@@ -72,6 +71,7 @@ import {
   ControlSectionSkeleton,
   SectionTitle,
 } from "@/configurator/components/chart-controls/section";
+import { BaseLayerSettings } from "@/configurator/components/chart-options-selector/base-layer-settings";
 import { ConversionUnitsField } from "@/configurator/components/chart-options-selector/conversion-units-field";
 import { CustomLayersSelector } from "@/configurator/components/custom-layers-selector";
 import {
@@ -595,7 +595,7 @@ const EncodingOptionsPanel = ({
       )}
       {/* FIXME: should be generic or shouldn't be a field at all */}
       {field === "baseLayer" && (
-        <ChartMapBaseLayerSettings chartConfig={chartConfig as MapConfig} />
+        <BaseLayerSettings chartConfig={chartConfig as MapConfig} />
       )}
       {field === "customLayers" && <CustomLayersSelector />}
       {encoding.sorting &&
@@ -2628,70 +2628,6 @@ const ChartImputation = ({ chartConfig }: { chartConfig: ChartConfig }) => {
           onChange={(e) => {
             updateImputationType(e.target.value as ImputationType);
           }}
-        />
-      </ControlSectionContent>
-    </ControlSection>
-  );
-};
-
-const ChartMapBaseLayerSettings = ({
-  chartConfig,
-}: {
-  chartConfig: MapConfig;
-}) => {
-  const locale = useLocale();
-  const [_, dispatch] = useConfiguratorState(isConfiguring);
-
-  useEffect(() => {
-    const map = getMap();
-
-    if (chartConfig.baseLayer.locked) {
-      if (map !== null) {
-        dispatch({
-          type: "CHART_FIELD_UPDATED",
-          value: {
-            locale,
-            field: null,
-            // FIXME: shouldn't be a field if not mapped to a component
-            path: "baseLayer.bbox",
-            value: map.getBounds().toArray(),
-          },
-        });
-      }
-    } else {
-      dispatch({
-        type: "CHART_FIELD_UPDATED",
-        value: {
-          locale,
-          field: null,
-          path: "baseLayer.bbox",
-          value: undefined,
-        },
-      });
-    }
-  }, [chartConfig.baseLayer.locked, dispatch, locale]);
-
-  return (
-    <ControlSection hideTopBorder>
-      <SectionTitle closable>
-        <Trans id="chart.map.layers.base">Base Layers</Trans>
-      </SectionTitle>
-      <ControlSectionContent gap="large">
-        <ChartOptionCheckboxField
-          label={t({
-            id: "chart.map.layers.base.show",
-            message: "Show",
-          })}
-          field={null}
-          path="baseLayer.show"
-        />
-        <ChartOptionSwitchField
-          label={t({
-            id: "chart.map.layers.base.view.locked",
-            message: "Locked view",
-          })}
-          field={null}
-          path="baseLayer.locked"
         />
       </ControlSectionContent>
     </ControlSection>
