+++ app/configurator/components/chart-options-selector/base-layer-settings.tsx
@@ -1,0 +1,84 @@
+/* eslint-disable react/jsx-no-undef */
+import { t, Trans } from "@lingui/macro";
+import { useEffect } from "react";
+
+import { getMap } from "@/charts/map/ref";
+import { MapConfig } from "@/config-types";
+import {
+  ControlSection,
+  ControlSectionContent,
+  SectionTitle,
+} from "@/configurator/components/chart-controls/section";
+import {
+  ChartOptionCheckboxField,
+  ChartOptionSwitchField,
+} from "@/configurator/components/field";
+import {
+  isConfiguring,
+  useConfiguratorState,
+} from "@/configurator/configurator-state";
+import { useLocale } from "@/locales/use-locale";
+
+export const BaseLayerSettings = ({
+  chartConfig,
+}: {
+  chartConfig: MapConfig;
+}) => {
+  const locale = useLocale();
+  const [_, dispatch] = useConfiguratorState(isConfiguring);
+
+  useEffect(() => {
+    const map = getMap();
+
+    if (chartConfig.baseLayer.locked) {
+      if (map !== null) {
+        dispatch({
+          type: "CHART_FIELD_UPDATED",
+          value: {
+            locale,
+            field: null,
+            // FIXME: shouldn't be a field if not mapped to a component
+            path: "baseLayer.bbox",
+            value: map.getBounds().toArray(),
+          },
+        });
+      }
+    } else {
+      dispatch({
+        type: "CHART_FIELD_UPDATED",
+        value: {
+          locale,
+          field: null,
+          path: "baseLayer.bbox",
+          value: undefined,
+        },
+      });
+    }
+  }, [chartConfig.baseLayer.locked, dispatch, locale]);
+
+  return (
+    <ControlSection hideTopBorder>
+      <SectionTitle closable>
+        <Trans id="chart.map.layers.base">Base Layers</Trans>
+      </SectionTitle>
+      <ControlSectionContent gap="large">
+        <ChartOptionCheckboxField
+          label={t({
+            id: "chart.map.layers.base.show",
+            message: "Show",
+          })}
+          field={null}
+          path="baseLayer.show"
+        />
+        <ChartOptionSwitchField
+          label={t({
+            id: "chart.map.layers.base.view.locked",
+            message: "Locked view",
+          })}
+          field={null}
+          path="baseLayer.locked"
+        />
+      </ControlSectionContent>
+    </ControlSection>
+  );
+};
